Remove unused send presenters and move send handlers to packages
