+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/send/SendModule.kt
@@ -4,10 +4,15 @@ import io.horizontalsystems.bankwallet.core.*
 import io.horizontalsystems.bankwallet.entities.CoinValue
 import io.horizontalsystems.bankwallet.entities.CurrencyValue
 import io.horizontalsystems.bankwallet.entities.Wallet
+import io.horizontalsystems.bankwallet.modules.send.binance.SendBinanceHandler
 import io.horizontalsystems.bankwallet.modules.send.binance.SendBinanceInteractor
+import io.horizontalsystems.bankwallet.modules.send.bitcoin.SendBitcoinHandler
 import io.horizontalsystems.bankwallet.modules.send.bitcoin.SendBitcoinInteractor
+import io.horizontalsystems.bankwallet.modules.send.dash.SendDashHandler
 import io.horizontalsystems.bankwallet.modules.send.dash.SendDashInteractor
+import io.horizontalsystems.bankwallet.modules.send.eos.SendEosHandler
 import io.horizontalsystems.bankwallet.modules.send.eos.SendEosInteractor
+import io.horizontalsystems.bankwallet.modules.send.ethereum.SendEthereumHandler
 import io.horizontalsystems.bankwallet.modules.send.ethereum.SendEthereumInteractor
 import io.horizontalsystems.bankwallet.modules.send.submodules.address.SendAddressModule
 import io.horizontalsystems.bankwallet.modules.send.submodules.amount.SendAmountModule
