+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/send/binance/SendBinanceHandler.kt
@@ -1,13 +1,14 @@
-package io.horizontalsystems.bankwallet.modules.send
+package io.horizontalsystems.bankwallet.modules.send.binance
 
+import io.horizontalsystems.bankwallet.modules.send.SendModule
 import io.horizontalsystems.bankwallet.modules.send.submodules.address.SendAddressModule
 import io.horizontalsystems.bankwallet.modules.send.submodules.amount.SendAmountModule
 import io.horizontalsystems.bankwallet.modules.send.submodules.fee.SendFeeModule
 import io.reactivex.Single
 import java.math.BigDecimal
 
-class SendEosHandler(private val interactor: SendModule.ISendEosInteractor,
-                     private val router: SendModule.IRouter) : SendModule.ISendHandler,
+class SendBinanceHandler(private val interactor: SendModule.ISendBinanceInteractor,
+                         private val router: SendModule.IRouter) : SendModule.ISendHandler,
         SendAmountModule.IAmountModuleDelegate,
         SendAddressModule.IAddressModuleDelegate {
 
@@ -23,10 +24,6 @@ class SendEosHandler(private val interactor: SendModule.ISendEosInteractor,
         }
     }
 
-    private fun syncAvailableBalance() {
-        amountModule.setAvailableBalance(interactor.availableBalance)
-    }
-
     // SendModule.ISendHandler
 
     override lateinit var amountModule: SendAmountModule.IAmountModule
@@ -38,6 +35,7 @@ class SendEosHandler(private val interactor: SendModule.ISendEosInteractor,
     override val inputItems: List<SendModule.Input> = listOf(
             SendModule.Input.Amount,
             SendModule.Input.Address,
+            SendModule.Input.Fee(false),
             SendModule.Input.ProceedButton)
 
     override lateinit var delegate: SendModule.ISendHandlerDelegate
@@ -51,7 +49,10 @@ class SendEosHandler(private val interactor: SendModule.ISendEosInteractor,
     }
 
     override fun onModulesDidLoad() {
-        syncAvailableBalance()
+        amountModule.setAvailableBalance(interactor.availableBalance)
+
+        feeModule.setFee(interactor.fee)
+        feeModule.setAvailableFeeBalance(interactor.availableBinanceBalance)
     }
 
     override fun onAddressScan(address: String) {
@@ -65,7 +66,7 @@ class SendEosHandler(private val interactor: SendModule.ISendEosInteractor,
     }
 
     override fun onChangeInputType(inputType: SendModule.InputType) {
-
+        feeModule.setInputType(inputType)
     }
 
     // SendAddressModule.IAddressModuleDelegate
@@ -79,7 +80,7 @@ class SendEosHandler(private val interactor: SendModule.ISendEosInteractor,
     }
 
     override fun onUpdateAmount(amount: BigDecimal) {
-        amountModule.setAvailableBalance(amount)
+        amountModule.setAmount(amount)
     }
 
     override fun scanQrCode() {
