+++ app/src/main/java/io/horizontalsystems/bankwallet/modules/send/dash/SendDashHandler.kt
@@ -1,17 +1,17 @@
-package io.horizontalsystems.bankwallet.modules.send
+package io.horizontalsystems.bankwallet.modules.send.dash
 
+import io.horizontalsystems.bankwallet.modules.send.SendModule
 import io.horizontalsystems.bankwallet.modules.send.submodules.address.SendAddressModule
 import io.horizontalsystems.bankwallet.modules.send.submodules.amount.SendAmountModule
 import io.horizontalsystems.bankwallet.modules.send.submodules.fee.SendFeeModule
 import io.reactivex.Single
 import java.math.BigDecimal
 
-class SendBitcoinHandler(private val interactor: SendModule.ISendBitcoinInteractor,
-                         private val router: SendModule.IRouter) : SendModule.ISendHandler,
-        SendModule.ISendBitcoinInteractorDelegate,
+class SendDashHandler(private val interactor: SendModule.ISendDashInteractor,
+                      private val router: SendModule.IRouter) : SendModule.ISendHandler,
+        SendModule.ISendDashInteractorDelegate,
         SendAmountModule.IAmountModuleDelegate,
-        SendAddressModule.IAddressModuleDelegate,
-        SendFeeModule.IFeeModuleDelegate {
+        SendAddressModule.IAddressModuleDelegate {
 
     private fun syncValidation() {
         try {
@@ -26,11 +26,11 @@ class SendBitcoinHandler(private val interactor: SendModule.ISendBitcoinInteract
     }
 
     private fun syncAvailableBalance() {
-        interactor.fetchAvailableBalance(feeModule.feeRate, addressModule.currentAddress)
+        interactor.fetchAvailableBalance(addressModule.currentAddress)
     }
 
     private fun syncFee() {
-        interactor.fetchFee(amountModule.coinAmount.value, feeModule.feeRate, addressModule.currentAddress)
+        interactor.fetchFee(amountModule.coinAmount.value, addressModule.currentAddress)
     }
 
     // SendModule.ISendHandler
@@ -46,7 +46,7 @@ class SendBitcoinHandler(private val interactor: SendModule.ISendBitcoinInteract
     override val inputItems: List<SendModule.Input> = listOf(
             SendModule.Input.Amount,
             SendModule.Input.Address,
-            SendModule.Input.Fee(true),
+            SendModule.Input.Fee(false),
             SendModule.Input.ProceedButton)
 
     override fun onModulesDidLoad() {
@@ -62,7 +62,7 @@ class SendBitcoinHandler(private val interactor: SendModule.ISendBitcoinInteract
     }
 
     override fun sendSingle(): Single<Unit> {
-        return interactor.send(amountModule.validAmount(), addressModule.validAddress(), feeModule.feeRate)
+        return interactor.send(amountModule.validAmount(), addressModule.validAddress())
     }
 
     // SendModule.ISendBitcoinInteractorDelegate
@@ -104,13 +104,6 @@ class SendBitcoinHandler(private val interactor: SendModule.ISendBitcoinInteract
 
     override fun scanQrCode() {
         router.scanQrCode()
-    }
-
-    // SendFeeModule.IFeeModuleDelegate
-
-    override fun onUpdateFeeRate(feeRate: Long) {
-        syncAvailableBalance()
-        syncFee()
     }
 
 }
